Removed discontinued work of cross-compiling ROS fuerte
After the release of ROS groovy, there is no strong reason to
continue to work on the cross-compilation of the previous ROS
version fuerte. Hence, this commit removes the initial
non-functioning recipe and the reference in the README file.

For cross-compiling ROS fuerte, the repository at
https://github.com/yida/yida-collections
might be a better starting point anyway.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>