Revert "simplified groovy installation on target"
In the current state, the catkin recipe does not work on the target system.
We want to keep the native compilation of ros groovy working, even if our
current development is focusing on the cross-compilation.
To make the native compilation of ros groovy more stable, we go back to the
previous state where even catkin was compiled on the target system.
Hence, this commit reverts commit 2c8274b74f263119f03953046c30155c74adac47
to obtain that previous state.

The conflicts in the files of this commit were manually resolved.