Revert "removing all roscpp_core packages from the created ros source archive"
The commit 83c813b8c32d4666c532e6622c6e2deb4772dc8e was due to an experiment
to cross-compile some packages and compile other on the target.
It turned out that there are a couple of difficulties and stopped working on
that experiment.
To obtain a working native compilation of ros-groovy,
this commit reverts the commit 83c813b8c32d4666c532e6622c6e2deb4772dc8e to
a state where the installation and compilation worked.