topic-tools: Add rostest to DEPENDS
The topic-tools fails during the do_compile step at:
.../sysroots/x86_64-linux/usr/lib/libboost_thread-mt.so: could not read symbols: File in wrong format

The failure comes from the rostest-native package which delivers a
header file with boost dependencies and therefore add the relevant
library to the link path. Add rostest to the DEPENDS to use the
header and libraries from the cross sysroot and keep the native
rostest for the binary.

Fixed #51

Signed-off-by: Stefan Herbrechtsmeier <stefan@herbrechtsmeier.net>