genmsg: removing runtime dependency after patch in oe-core (resolves #54)
The recipe python-pprint rdepends on python-io. As this was not reflected in
the oe-core repository, we added the dependency in genmsg.
Now after the commit abe7bf9992e298f1b53e790eee7b064a9e4e8589 in oe-core,
we can remove the runtime dependency in genmsg.

This commit resolves issue #54.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>