console-bridge: no additional files in dev package
Currently, it is not clear if the additional files are required in
the normal package or the dev package. Hence, we do not touch them
(they are not added to any package or deleted after the installation)
until the right operation is clear.

Until then, the console-bridge emits the warning
WARNING: QA Issue: console-bridge: Files/directories were installed but not shipped
  /usr/share
  /usr/share/console_bridge
  /usr/share/console_bridge/package.xml
  /usr/share/console_bridge/cmake
  /usr/share/console_bridge/cmake/console_bridge-config.cmake