roscpp: removing roscpp's own routine to find pthread (resolves #125)
When updating from 1.9.41 to 1.9.47, the commit
https://github.com/ros/ros_comm/commit/c5dfa6305e883f76dc2aa52d27d2977e3d66ea2e
adds a routine in the CMakeLists.txt to find the pthread library.
This routine is not suitable for the use with our tool chain, i.e.,
it reports an error (cf. #125) when using roscpp-native.

This commit adds a patch that reverts that change in roscpp's
CMakeLists.txt.