check-patch-files.sh: initial commit
The check-patch-files script checks that all patch files in the
given directory are annotated with an 'Upstream-Status', and
that all patch files are at least noted in their top-level
directory (so we can assume they are used).

In case a violation is detected, it returns a non-zero exit code,
so that a travis-ci reports a violation as build failure.