packagegroup-ros-world: removing urdfdom-headers (re-addressing #177).
The urdfdom-headers package was removed from packagegroup-ros-world
in #182, because it is an empty package and causes the image
creation to fail (cf. #177).
In commit 9ff8d1f7e9130dca1455651bd5d58875fc95c30e, urdfdom-headers
was added again while adding new ROS packages to
packagegroup-ros-world. As part of the manual regression testing,
executing 'bitbake packagegroup-ros-world' for the different qemu
architectures, but not building full images, this problem has
slipped through the regression testing the last few months. This
issue was reported on the meta-ros mailing list by Parag
Batavia [1].

[1] https://groups.google.com/forum/#!topic/meta-ros/VZo3IUsI_48