octomap: simplifying recipe
When I investigated the issue #291 with Kristof's patch applied,
I noticed that `bitbake octomap` failed.

Due to my inspection, I believe that the inheritance on ros
was only needed as the ROS_SPN and ROS_SP variables were used in
the recipe. After simply using the default variables, BPN and BP,
I removed the inheritance on ros.

Furthermore, as meta-ros only has one recipe for octomap base
library for now, we do not need to split the recipe definition
in an include and a recipe file.
Probably, this was done as premature optimization believing
that the libraries from the octomap repository, e.g., octovis,
would be added shortly after.

Testing with `bitbake packagegroup-ros-world` reported no errors.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>