ar-track-alvar: updating to 0.4.2
With the poky-dizzy distribution, the do_rootfs task for
core-image-ros-world fails with:

  ERROR: Unable to install packages. Command '/[...]/build/tmp/sysroots/x86_64-linux/usr/bin/smart --quiet --data-dir=/[...]/core-image-ros-world/1.0-r0/rootfs/var/lib/smart install -y run-postinsts@all packagegroup-core-boot@beaglebone packagegroup-ros-world@all' returned 1:
  error: Can't install ar-track-alvar-0.4.1-r0@cortexa8hf_vfp_neon: no package provides libmedianFilter.so

Build Configuration: poky 1.7.1;
  poky: "dizzy:ec75238f6cc2d2d8d40e0268f6d2acc070cbe9a4";
  meta-openembedded: "dizzy:9efaed99125b1c4324663d9a1b2d3319c74e7278"

To resolve this problem, this commit updates ar-track-alvar to the
latest Hydro version 0.4.2. Unfortunately, there is no archive file
for version 0.4.2, so the recipe uses the git repository with the
commit intended to mark version 0.4.2 to fetch the source code.
Due to the update, this commit also removes the upstream-accepted
patch file from this repository here.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>