frontier-exploration: point to license provided in package.xml
The frontier-exploration recipe stated that the software is under
BSD license, probably due to a copy-and-paste inattention. The
commit now points to the license line for version 0.2.2 and changes
the license to GPLv3, which is stated in the package.xml.
Strangely, the package.xml states that the frontier_exploration
ROS package is licensed under GPLv3, but the license text in the
LICENSE file is the GPLv2.1 terms and conditions; so the actual
intended license by the copyright holders remain unclear.
However, assuming the conditions for GPLv3 must be fulfilled
for usage and distribution is a 'safe' approximation, even if
the conditions for GPLv2.1 apply.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>