images: Remove spurious post-processing command
The 'remove_packaging_data_files' post-processing command generates a
warning for Poky version 1.5+ and is no longer needed.  Section 24.4.9
of the Yocto manual (v1.8) explains that these files are automatically
removed for images not using the package-management feature.

This has been maintained for backwards-compatibility but as of the 0.2
release of this layer, Poky version 1.4 is several years old so we can
reasonably remove this post-processing command (discussion [1]).

[1] https://www.mail-archive.com/meta-ros%40googlegroups.com/msg00086.html

Signed-off-by: Ash Charles <ashcharles@gmail.com>