pcl: updating to 1.7.2
The update to the latest version 1.7.2 was triggered by compilation
errors in pcl 1.7.0 (i.e., in file common/include/pcl/point_types.h)
due to the update of boost to 1.60.0 in openembedded-core [1].
This failure was detected by the internal CI build on 2016-01-15.

Due to the version update, the commit drops patch
`Fix incompatibility with Boost 1.56`, as the change is included
upstream in version 1.7.2. This commit also adds the upstream-accepted
patch `Add "WITH_OPENGL" option (...)` to make the dependency to OpenGL
libraries optional.

The commit configures package dependencies and compiles with the
least possible packages, so that the pcl-ros recipe compiles. As a
result, all optional packages in PCL, except qhull, can be set to FALSE.

If qhull is deactivated, the do_configure task of pcl-ros reports:

  | -- Could NOT find qhull (missing:  QHULL_LIBRARY)
  | ** WARNING ** surface features related to qhull will be disabled

However, in the do_compile task of pcl-ros, the build process assumes
that the surface features related to qhull are actually never disabled,
i.e., if qhull is deactivated, compiling pcl-ros fails with:

  | In file included from [...]/pcl_ros/src/pcl_ros/surface/convex_hull.cpp:40:0:
  | [...]/pcl_ros/include/pcl_ros/surface/convex_hull.h:44:37: fatal error: pcl/surface/convex_hull.h: No such file or directory
  | compilation terminated.
  | CMakeFiles/pcl_ros_surface.dir/build.make:89: recipe for target 'CMakeFiles/pcl_ros_surface.dir/src/pcl_ros/surface/convex_hull.cpp.o' failed
  | make[2]: *** [CMakeFiles/pcl_ros_surface.dir/src/pcl_ros/surface/convex_hull.cpp.o] Error 1
  | make[2]: Leaving directory '[...]/build'
  | CMakeFiles/Makefile2:1173: recipe for target 'CMakeFiles/pcl_ros_surface.dir/all' failed
  | make[1]: *** [CMakeFiles/pcl_ros_surface.dir/all] Error 2

[1] http://cgit.openembedded.org/openembedded-core/commit/?id=17d77db5e1203e18cdd178e7d9a457310425ebdc

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>