urdfdom-headers: Fix missing package
Bitbake gives a dependency error when IMAGE_FEATURES += " dev-pkgs" is set.

This is because the urdfdom-headers package contains headers only, as a
result bitbake packages them into urdfdom-headers-dev and leaves
urdfdom-headers empty.

Since there are no files to be packaged into urdfdom-headers bitbake
doesn't create the package. urdfdom-headers is a dependency
of urdfdom-headers-dev by default so bitbake errors when building an image
with dev-pkgs enabled.

ALLOW_EMPTY tells bitbake to create a package even when nothing is in it.
This resolves the dependency issue of urdfdom-headers-dev
http://www.embeddedlinux.org.cn/OEManual/recipes_packages.html