python-six: remove recipe already provided by others
The Openembedded Layer Index [1] shows that python-six in
version 1.10.0 is already provided in the openembedded-core layer
and in the meta-python layer, so there is no need for a third copy
of the python-six recipe in meta-ros.

The recipe python-six is provided since commit a36869c7 in the
meta-openembedded repository (commit date: 2015-04-13), and
is provided since commit 9979e35f in the openembedded-core
repository (commit date: 2016-07-20).

In my opinion, at least the commit a36869c7 is far enough in the
past, so that we can safely assume that all new meta-ros use a setup
where python-six is included at least in meta-python.

For users of older versions, we include a remark in the README.md to
allow `git blame` to pinpoint to this commit here.

[1] https://layers.openembedded.org/layerindex/branch/master/recipes/?q=python-six
[2] http://cgit.openembedded.org/meta-openembedded/commit/?id=a36869c700bed940f7f0aa4b9703ae630dc84eac
[3] http://cgit.openembedded.org/openembedded-core/commit/?id=9979e35ffb0dfb20d8aaffa8e2ec5e33f27fc009

Signed-off-by: Lukas.Bulwahn <lukas.bulwahn@oss.bmw-carit.de>