poco: use meta-oe's recipe instead of own one
The meta-oe provides since commit 8700ba38@openembedded/meta-openembedded
(commit date: 2016-01-14) [1] a recipe for the POCO C++ Library.
Consequently, this commit removes the libpoco recipe in the
meta-ros layer.

As meta-ros already depends on the meta-oe layer, the poco
recipe is available in release version since 2016-01-14,
and the poco recipe in meta-ros is redundant for these versions.
However, this commit still provides the currently latest poco
recipe [2] for users that are not using a recent version of the
meta-oe layer.

[1] http://cgit.openembedded.org/meta-openembedded/commit?id=8700ba38804af3c27f3662737f679afa1bdc86da
[2] http://cgit.openembedded.org/meta-openembedded/tree/meta-oe/recipes-support/poco?id=0a2bd4f9784253a8a8dda35e3c4dbd75931d3564

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>