diagnostics: updating to 1.8.10 and adjusting dependencies
In 2014, the update of the diagnostics recipes from 1.8.3 to 1.8.4
failed due to reasons that seem not to be relevant anymore in the
current version 1.8.10. As the diagnostic-aggregator recipe failed
with gcc6 (cf. https://github.com/bmwcarit/meta-ros/issues/392),
I revisited the diagnostics recipes and this commit updates them.
The current diagnostics recipes in version 1.8.10 do not fail with
gcc6, and is one step forward to build meta-ros with gcc6.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>