python-imaging: removing recipe maintained in meta-python
On 2016-06-08, the commit 000095f3 [1, 2] added the python-imaging
recipe that was dropped in OpenEmbedded Core in commit b03ebfaee [3]
on 2016-06-02.

Coincidentally, python-imaging was also added in meta-python
on 2016-06-08 [4], and it has been further maintained in that layer [5].

Instead of continuing to maintain the recipe in meta-ros, we decide to
rely on the recipe provided in meta-python, and hence, remove the
recipe in meta-ros.

[1] https://github.com/bmwcarit/meta-ros/commit/000095f3f4d5f6d15ffab11ebe4d33ad90541967
[2] https://github.com/bmwcarit/meta-ros/issues/393
[3] http://cgit.openembedded.org/openembedded-core/commit/meta/recipes-devtools/python?id=b03ebfaee278eb98b4bd0e11a682ac0cda192086
[4] http://cgit.openembedded.org/meta-openembedded/commit/meta-python/recipes-devtools/python/python-imaging_1.1.7.bb?id=56a4d21706e0219bbe488c2bd1450247f4f89d0d
[5] http://cgit.openembedded.org/meta-openembedded/tree/meta-python/recipes-devtools/python/python-imaging_1.1.7.bb

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>