python-rosdistro: add missing lib rosdep2 depends on
roswtf reports uninitialized /etc/ros structure and recommends
to fix the issue by running `rosdep init`. But rosdep throws an
exception when trying to import the missing rosdistro library.

The library is supposed to be dropped in kinetic, but it's still
needed in indigo which the layer is based on at the moment.
So, the patch adds the library to the layer.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>