ros-mavlink: add recipe
A patch for the config file is also necessary because the include directory
path was being hardcoded in the generated file, which caused problems for cross
compilation. That patch has already been applied on upstream but for a newer
version, so we're backporting it here.

Apparently, the Kinetic release for this package is supposed to work fine with
indigo distribution. That could be tried later, so that we can get rid of the
local patch.

Authors: JochiPochi <john.aleman@cyphyworks.com>
         Gustavo Jose de Sousa <gustavo.sousa@intel.com>