eigen-stl-containers: depend on libeigen
Without the dependency on libeigen, `bitbake eigen-stl-containers`
can possibly fail with:

```
| CMake Error at CMakeLists.txt:8 (find_package):
|   By not providing "FindEigen3.cmake" in CMAKE_MODULE_PATH this project has
|   asked CMake to find a package configuration file provided by "Eigen3", but
|   CMake did not find one.
|
|   Could not find a package configuration file provided by "Eigen3" with any
|   of the following names:
|
|     Eigen3Config.cmake
|     eigen3-config.cmake
|
|   Add the installation prefix of "Eigen3" to CMAKE_PREFIX_PATH or set
|   "Eigen3_DIR" to a directory containing one of the above files.  If "Eigen3"
|   provides a separate development package or SDK, be sure it has been
|   installed.
|
|
| -- Configuring incomplete, errors occurred!
```

The failure only occurs if libeigen has not been installed
before eigen-stl-containers is configured. Hence, the regular
regression testing with `bitbake core-image-ros-world`,
which builds many packages in parallel, did not uncover this
because libeigen was usually installed before
eigen-stl-containers was configured.

However, the issue is clearly reproducible with
`bitbake eigen-stl-containers libeigen -c cleanall && bitbake eigen-stl-containers`

The missing dependency was probably overlooked in the creation of
the eigen-stl-containers recipe, i.e., in
commit a255e67c9e5eac62c537ace6d76c885d5b0ae11b.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>