orocos-kdl: update to git version to avoid #469
This commit updates the Orocos Kinematics and Dynamics C++ library
recipes to the current git commit. On this current git commit, the
issue #469 does not occur anymore, and hence, we do not have to
address the issue anymore.

My regression testing also did not show any build problems when
removing the do_install_append function in the orocos-kdl recipe.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>