poco: drop recipe (resolves #377)
The poco recipe is in meta-oe now for quite some time and is
actively updated and maintained there. Also, meta-ros with its
recent changes, especially the changes due to recipe-specific
sysroots, is unlikely to build with older versions of meta-oe.
Hence, this commit drops the poco recipe in meta-ros for good.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>