moveit-ros-planning: call python3 for dynamic reconfigurations
If python2 is available in build environment then according to
PEP-394 python should reference python2. But in case of building
a python3 flavor of meta-ros PYTHONPATH is set explicitly to
python3 libraries. When python2 is called with PYTHONPATH
pointing to python3 libs the following error triggers:

    Could not find platform independent libraries <prefix>
    Could not find platform dependent libraries <exec_prefix>
    Consider setting $PYTHONHOME to <prefix>[:<exec_prefix>]

This patch makes the dynamic reconfigurations call python3
explicitly in case a python3 flavor of meta-ros is being
built.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>