robot-model: move some packages to own repositories
The recipes collada-parser, collada-urdf and kdl-parser have been
moved to new repositories and are further developed there. This
commit updates the recipes to the current versions in the new
repositories. Furthermore, it adds the new recipe kdl-parser-py,
but this recipe is still deactivated in packagegroup-ros-world, as
python-orocos-kdl is not yet properly building.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>