yaml-cpp: disable test & util
In a specific build setup that we are not able to reproduce anymore,
building yaml-cpp failed when compiling the utilities and tests. Rather
than addressing the issue in yaml-cpp, this commit deactivates the tests
and utilities that failed with that setup. As both the tests and utilities
are not installed and not packaged, this commit should not cause any build
failures for any other unrelated setup.