geometry: updating to 1.11.9
In the tf recipe, this commit adjusts the license-file line and
drops the upstream-accepted patch due to update.

With the update of eigen-conversions and tf-conversions,
the compilation of both packages fails with:
```
| [...]: fatal error: Eigen/Core: No such file or directory
|  #include <Eigen/Core>
|           ^~~~~~~~~~~~
| compilation terminated.
```
It fails as the upstream development uses the Eigen3 CMake
find script instead of the ROS cmake_modules find script,
and the Eigen3 CMake find script does not work in the current
openembedded-core and meta-openembedded setup (cf. #501).

Hence, this commit applies a patch to revert the CMakeLists.txt
to use ROS cmake_modules to find Eigen. This patch must be applied
until issue #501 is resolved in the upstream development.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>