New plugin for generating catkin recipes
Currently the plugin only allow for processing a single
ROS package in a repository. If a repository contains more than
one package, or the package is not in the root of the repository,
then use the `--src-subdir=<dir>` option.

ROS repositories generally do not use `master` as their default
branch, so be sure to include the correct branch for the desired
distribution as part of the URI: `<URI>;branch=indigo-devel`