Fixed DEPENDS for dynamic-reconfigure recipe
rospkg is necessary to generate dreconf files. dynamic-reconfigure was not
providing proper build-time depencency to packages that use it.

This patch series adds ${PYTHON_PN}-roskpg to DEPENDS of dynamic-reconfigure.bb
and removes unnecessary reference to it in all recipes that are using dynamic-reconfigure.

There were a fix for theora-image-transport with missing libvorbis as well.