wxwidgets: Change where wx-config is installed
Users of wxwidgets will want to execute what the symlink wx-config
points to, but it's installed under ${bindir}, which isn't included in
the sysroot (because it contains wxrc, a cross-compiled executable). So
instead, place it under ${libdir}/wx/config and require that users of
wxwidgets cause it to be invoked by adding passing
"WX_CONFIG=${RECIPE_SYSROOT}${libdir}/wx/config/wx-config" when
configuring.