ros-distro-platform-preferred-versions.inc: Drop PREFERRED_VERSION_linux-*
raspberrypi4 fails to build with kernel v4.14 when the "nodistro" config
is used:

  with warrior:
  make[3]: *** No rule to make target
  'arch/arm/boot/dts/bcm2708-rpi-zero-w.dtb'.  Stop.
  arch/arm/Makefile:343: recipe for target 'bcm2708-rpi-zero-w.dtb' failed

  with zeus (including meta-updater):
  make[3]: *** No rule to make target
  'arch/arm/boot/dts/bcm2711-rpi-4-b.dtb'.  Stop.
  arch/arm/Makefile:343: recipe for target 'bcm2711-rpi-4-b.dtb' failed

Therefore, as of warrior, change the policy for selecting the version of
the kernel to build: instead of choosing it based on that provided by
ROS_DISTRO_BASELINE_PLATFORM (currently "ubuntu-bionic"), use the
default version provided by the BSP layer for the OpenEmbedded release
series being used => drop setting PREFERRED_VERSION_linux-* in this file.

Note that webOS OSE sets the preferred version to whatever works with
it.