{melodic} semantic-point-annotator, robot-self-filter, seed-smartactuator-sdk: fix build with boost 1.69
Deprecated signals library was dropped in 1.69.0 release:
https://www.boost.org/users/history/version_1_69_0.html

Signals (v1) is now removed. Its removal was announced in 1.68 and its
deprecation was announced in 1.54. Boost 1.68 is the last release that
provides this library. Users are encouraged to use Signals2 instead. The
Boost community thanks Douglas Gregor for his work on Signals which
served its users well and which also inspired Signals2.

Signed-off-by: Martin Jansa <martin.jansa@lge.com>