{galactic} libphidget22-upstream: disable debian package names
Resolve problem with libphidget22 and Debian package naming rules.

The Debian package naming convention will rename a package after the
shared library name if it is the only thing present in the package.

This resulted in libphidget22-upstream being renamed to libphidget22
and there being 2 sets of packages present with that name.

As a workaround we set AUTO_LIBNAME_PKGS to empty to disable the Debian
naming rules for that package.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>
Signed-off-by: Martin Jansa <martin.jansa@lge.com>