{galactic} openvslam: Retain CMAKE_CXX_FLAGS from OpenEmbedded
The CMakeLists.txt for FBoW tries to automatically set the compiler
options by setting CMAKE_CXX_FLAGS.  This interferes with the compiler
options set by the build system.  It causes FBoW to fail to find
iostream because --sysroot is not set.

We patch this part of FBoW CMakeLists.txt to prepend the existing
value when setting the variable.  This allows openvslam to build
for both qemux86-64 and raspberrypi4-64.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>
Signed-off-by: Martin Jansa <martin.jansa@lge.com>