Use CMake to build libyaml
The CMake files are present in the libyaml repository but not packaged in
releases.

Using CMake instead of autotools makes it possible for libyaml-vendor to
export dependencies instead of exporting libraries.  This prevents host
contamination from the absolute paths to libyaml being included in the
libyaml-vendor CMake files.

Ubuntu is still using autotools, but there are projects that use the
CMake build system for libyaml.  This includes the CI/CD build system for
the  C Client Library for Kubernetes.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>