libyaml:  Add a custom recipe for libyaml
Packages that depend on libyaml can use either pkgconfig or CMake to
find it.  When pkgconfig is used, it results in the absolute path for
the host build machine to be embedded in CMake files for the target.

The upstream libyaml repository has CMake files that don't get published
in the release tarball or used by the oe-core recipe.

We are patching the CMake files from upstream back in so that we may
use it instead of autotools and pkgconfig.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>